requirements.txt
websocket-client
selenium

